requirements.txt
aiocache==0.12.3
alembic==1.18.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==26.1.0
blinker==1.9.0
cachelib==0.13.0
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
flasgger==0.9.7.1
Flask==3.1.3
Flask-Caching==2.3.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
mistune==3.2.0
mysql-connector-python==9.6.0
packaging==26.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
SQLAlchemy==2.0.47
SQLAlchemy-Utils==0.42.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.6
